separated home page and new playlist page
