--- src/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      WRONG App
-    </div>
-  );
-}
-
-export default App;
